refactor(cli): extract --setup into `acpctl gateway setup` subcommand
Move the gateway setup logic from a --setup flag on `acpctl get gateway`
into a dedicated `acpctl gateway setup <name>` subcommand. Get commands
should be read-only; setup mutates local state (extracts certs, starts
port-forward, registers with openshell CLI).

The internals of setupOpenshellGateway are unchanged — only the
registration path moved. Added a TODO for future Route/Ingress support.
